+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVM4NY4J88SZW5BKXAAG4TGZ"
+        "01JVMB8PHEDJXX9TZ48Y3XVZ1K"
     )
     r"""The id of the resource"""